IAuthClient now throws a checked exception for a failed login and passes through RuntimeExceptions
*** This is likely an incompatible change for applications directly using the IAuthClient interface ***
Previously users of IAuthClient.getNewToken() need to catch a NotAuthorizedException(a RuntimeException)
in order to detect a failed login.  Since this is a non-ignorable exception, this is now a checked
exception (LoginExcpetion).
Also, other types of RuntimeExceptions that happen during the attempt to
validate the user against a remote service are now passed through to the caller
of getNewToken().  Previously they were all swallowed by the call.
This will result in some RuntimeErrors being passed higher up the chain in some
apps.  Existing tests had to be changed because a bad url now throws an
NotFoundException instead of just returning a null token.
***

* Improved docs for getNewToken()
* Downgraded the surefire plugin from 2.19.1 to 2.18.1 to bypass a bug in NetBeans
where 2.19.1 will not show the JUnit test window:
https://netbeans.org/bugzilla/show_bug.cgi?id=257563